server/requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
motor>=3.3.1
python-dotenv>=1.0.0
